requirements.txt
zai-sdk
edge-tts
python-dotenv
pydub